refactor: tighten feature folder boundaries
