+++ src/app/settings/switch-scanning/switch-input-presentation.ts
