+++ src/app/settings/switch-scanning/switch-input-setup.tsx
