+++ src/app/settings/switch-scanning/switch-scanning-advanced-settings.tsx
