+++ src/app/settings/switch-scanning/switch-scanning-settings.test.tsx
