+++ src/features/board/keyboard/use-board-keyboard.test.tsx
@@ -2,7 +2,10 @@ import { stubAudio } from "@shared/testing/stub-audio";
 import { preventSpeechEnd, stubSpeech } from "@shared/testing/stub-speech";
 import { beforeEach, describe, expect, test, vi } from "vitest";
 import { userEvent } from "vitest/browser";
-import { renderBoardViewer, TWO_BUTTON_BOARD } from "../test-utils";
+import {
+  renderBoardViewer,
+  TWO_BUTTON_BOARD,
+} from "../testing/render-board-viewer";
 
 describe("board keyboard shortcuts", () => {
   let speech: ReturnType<typeof stubSpeech>;
