+++ src/features/board/scanning/use-board-scanning.ts
